work towards a dynamic memory allocator for opencl

+++ swat-bridge/common.h
@@ -11,7 +11,7 @@ inline void safe_write(int fd, const void *buf, size_t count) {
         fprintf(stderr, "Write failed\n");
         perror("write");
         exit(1);
-    } else if (written != count) {
+    } else if ((size_t)written != count) {
         fprintf(stderr, "Expected a write of %lu bytes but actually wrote %lu "
                 "bytes\n", count, written);
         exit(1);
@@ -23,7 +23,7 @@ inline void safe_read(int fd, void *buf, size_t count) {
         fprintf(stderr, "Read failed\n");
         perror("read");
         exit(1);
-    } else if (r != count) {
+    } else if ((size_t)r != count) {
         fprintf(stderr, "Expected a read of %lu bytes but actually read %lu "
                 "bytes\n", count, r);
         exit(1);
